Mark current feed as read only after asking for the next one
This change enables the user to come back to the same feed if he has no time to read it at the current session or if he moves to another page.

The drawback is that there is a redirection when asking for the next feed (again). I need to investigate if it is possible to pass parameters to arbitrary urls in Rails, since the only way I know to do that is through named paths. And since there is no way of knowing if the next page is a feed or the topics/welcome page, I'll use the generic next_path for the time being.